src/cloud_functions/places_transformer/requirements.txt
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
pandas==2.0.0
numpy==1.24.2
functions-framework==3.4.0
blinker==1.6.2
click==8.1.6
cloudevents==1.9.0
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
watchdog==3.0.0
Werkzeug==2.3.6
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
python-dotenv==1.0.0
google-cloud-secret-manager==2.16.2
grpc-google-iam-v1==0.12.6